+++ p/insa/index.js
@@ -1,6 +1,20 @@
 'use strict'
 
+const _createParseLine = require('../../parse/line')
 const products = require('./products')
+const createParseBitmask = require('../../parse/products-bitmask')
+const createFormatBitmask = require('../../format/products-bitmask')
+
+const defaultProducts = {
+	nationalExp: true,
+	national: true,
+	regional: true,
+	suburban: true,
+	bus: true,
+	tram: true,
+	tourismTrain: true,
+}
+
 
 const transformReqBody = (body) => {
 	body.client = {
@@ -17,13 +31,49 @@ const transformReqBody = (body) => {
 	return body
 }
 
+const createParseLine = (profile, operators) => {
+	const parseLine = _createParseLine(profile, operators)
+
+	const parseLineWithMode = (l) => {
+		const res = parseLine(l)
+
+		res.mode = res.product = null
+		if ('class' in res) {
+			const data = products.bitmasks[parseInt(res.class)]
+			if (data) {
+				res.mode = data.mode
+				res.product = data.product
+			}
+		}
+
+		return res
+	}
+	return parseLineWithMode
+}
+
+const formatProducts = (products) => {
+	products = Object.assign(Object.create(null), defaultProducts, products)
+	return {
+		type: 'PROD',
+		mode: 'INC',
+		value: formatBitmask(products) + ''
+	}
+}
+
+const formatBitmask = createFormatBitmask(products)
+
+
 const insaProfile = {
 	locale: 'de-DE',
 	timezone: 'Europe/Berlin',
-	endpoint: 'http://reiseauskunft.insa.de/bin/mgate.exe',
+	endpoint: 'https://reiseauskunft.insa.de/bin/mgate.exe',
 	transformReqBody,
 
-	products: products,
+	products: products.allProducts,
+	parseProducts: createParseBitmask(products.allProducts, defaultProducts),
+	formatProducts,
+
+	parseLine: createParseLine,
 
 	journeyLeg: true,
 	radar: true