+++ p/oebb/index.js
@@ -3,6 +3,9 @@
 // todo: https://gist.github.com/anonymous/a5fc856bc80ae7364721943243f934f4#file-haf_config_base-properties-L5
 // todo: https://gist.github.com/anonymous/a5fc856bc80ae7364721943243f934f4#file-haf_config_base-properties-L47-L234
 
+const createParseBitmask = require('../../parse/products-bitmask')
+const createFormatBitmask = require('../../format/products-bitmask')
+const _createParseLine = require('../../parse/line')
 const _parseLocation = require('../../parse/location')
 const _createParseMovement = require('../../parse/movement')
 
@@ -23,6 +26,26 @@ const transformReqBody = (body) => {
 	body.lang = 'de'
 
 	return body
+}
+
+const createParseLine = (profile, operators) => {
+	const parseLine = _createParseLine(profile, operators)
+
+	const parseLineWithMode = (l) => {
+		const res = parseLine(l)
+
+		res.mode = res.product = null
+		if ('class' in res) {
+			const data = products.bitmasks[parseInt(res.class)]
+			if (data) {
+				res.mode = data.mode
+				res.product = data.product
+			}
+		}
+
+		return res
+	}
+	return parseLineWithMode
 }
 
 const parseLocation = (profile, l, lines) => {
@@ -59,6 +82,28 @@ const createParseMovement = (profile, locations, lines, remarks) => {
 	return parseMovement
 }
 
+const defaultProducts = {
+	nationalExp: true,
+	national: true,
+	interregional: true,
+	regional: true,
+	suburban: true,
+	bus: true,
+	ferry: true,
+	subway: true,
+	tram: true,
+	onCall: true
+}
+const formatBitmask = createFormatBitmask(products)
+const formatProducts = (products) => {
+	products = Object.assign(Object.create(null), defaultProducts, products)
+	return {
+		type: 'PROD',
+		mode: 'INC',
+		value: formatBitmask(products) + ''
+	}
+}
+
 const oebbProfile = {
 	locale: 'de-AT',
 	timezone: 'Europe/Vienna',
@@ -66,10 +111,14 @@ const oebbProfile = {
 	endpoint: 'http://fahrplan.oebb.at/bin/mgate.exe',
 	transformReqBody,
 
-	products: products,
+	products: products.allProducts,
 
+	parseProducts: createParseBitmask(products.allProducts, defaultProducts),
+	parseLine: createParseLine,
 	parseLocation,
 	parseMovement: createParseMovement,
+
+	formatProducts,
 
 	journeyLeg: true,
 	radar: true