+++ p/vbb/index.js
@@ -11,8 +11,12 @@ const _parseLocation = require('../../parse/location')
 const _createParseJourney = require('../../parse/journey')
 const _createParseDeparture = require('../../parse/departure')
 const _formatStation = require('../../format/station')
+const createParseBitmask = require('../../parse/products-bitmask')
+const createFormatBitmask = require('../../format/products-bitmask')
 
-const products = require('./products')
+const modes = require('./modes')
+
+const formatBitmask = createFormatBitmask(modes)
 
 const transformReqBody = (body) => {
 	body.client = {type: 'IPA', id: 'VBB', name: 'vbbPROD', v: '4010300'}
@@ -26,11 +30,19 @@ const transformReqBody = (body) => {
 const createParseLine = (profile, operators) => {
 	const parseLine = _createParseLine(profile, operators)
 
-	const parseLineWithMoreDetails = (l) => {
+	const parseLineWithMode = (l) => {
 		const res = parseLine(l)
 
-		res.name = l.name.replace(/^(bus|tram)\s+/i, '')
-		const details = parseLineName(res.name)
+		res.mode = res.product = null
+		if ('class' in res) {
+			const data = modes.bitmasks[parseInt(res.class)]
+			if (data) {
+				res.mode = data.mode
+				res.product = data.product
+			}
+		}
+
+		const details = parseLineName(l.name)
 		res.symbol = details.symbol
 		res.nr = details.nr
 		res.metro = details.metro
@@ -39,7 +51,7 @@ const createParseLine = (profile, operators) => {
 
 		return res
 	}
-	return parseLineWithMoreDetails
+	return parseLineWithMode
 }
 
 const parseLocation = (profile, l, lines) => {
@@ -119,6 +131,24 @@ const formatStation = (id) => {
 	return _formatStation(id)
 }
 
+const defaultProducts = {
+	suburban: true,
+	subway: true,
+	tram: true,
+	bus: true,
+	ferry: true,
+	express: true,
+	regional: true
+}
+const formatProducts = (products) => {
+	products = Object.assign(Object.create(null), defaultProducts, products)
+	return {
+		type: 'PROD',
+		mode: 'INC',
+		value: formatBitmask(products) + ''
+	}
+}
+
 const vbbProfile = {
 	locale: 'de-DE',
 	timezone: 'Europe/Berlin',
@@ -131,15 +161,17 @@ const vbbProfile = {
 
 	transformReqBody,
 
-	products: products,
+	products: modes.allProducts,
 
 	parseStationName: shorten,
 	parseLocation,
 	parseLine: createParseLine,
+	parseProducts: createParseBitmask(modes.allProducts, defaultProducts),
 	parseJourney: createParseJourney,
 	parseDeparture: createParseDeparture,
 
 	formatStation,
+	formatProducts,
 
 	journeysNumF: false,
 	journeyLeg: true,