+++ parse/line.js
@@ -2,14 +2,8 @@
 
 const slugg = require('slugg')
 
+// todo: are p.number and p.line ever different?
 const createParseLine = (profile, operators) => {
-	const byBitmask = []
-	for (let product of profile.products) {
-		for (let bitmask of product.bitmasks) {
-			byBitmask[bitmask] = product
-		}
-	}
-
 	const parseLine = (p) => {
 		if (!p) return null // todo: handle this upstream
 		const res = {
@@ -19,9 +13,8 @@ const createParseLine = (profile, operators) => {
 			public: true
 		}
 		// todo: what is p.prodCtx && p.prodCtx.num?
-		// todo: what is p.number?
 
-		// This is terrible, but FPTF demands an ID. Let's pray for HaCon to expose an ID.
+		// This is terrible, but FPTF demands an ID. Let's pray for VBB to expose an ID.
 		// todo: find a better way
 		if (p.line) res.id = slugg(p.line.trim())
 		else if (p.name) res.id = slugg(p.name.trim())
@@ -31,12 +24,7 @@ const createParseLine = (profile, operators) => {
 			res.productCode = +p.prodCtx.catCode
 		}
 
-		if ('class' in res) {
-			// todo: what if `res.class` is the sum of two bitmasks?
-			const product = byBitmask[parseInt(res.class)]
-			res.mode = product && product.mode || null
-			res.product = product && product.id || null
-		}
+		// todo: parse mode, remove from profiles
 
 		if ('number' === typeof p.oprX) {
 			res.operator = operators[p.oprX] || null